general: Use g_memdup2
Also make sure that all of the parameters are gsize
which already matches what we get (strlen returns
size_t). g_memdup took a guint.

+++ src/nautilus-file-operations.h
@@ -71,7 +71,7 @@ void nautilus_file_operations_new_file    (GtkWidget                 *parent_vie
 					   const char                *parent_dir,
 					   const char                *target_filename,
 					   const char                *initial_contents,
-					   int                        length,
+					   gsize                      length,
 					   NautilusCreateCallback     done_callback,
 					   gpointer                   data);
 void nautilus_file_operations_new_file_from_template (GtkWidget               *parent_view,
